fr this time

+++ search.py
@@ -29,8 +29,9 @@ class Search:
 			self.pca_to_samples[self.texel_embeddings[idx]] = sample
 		
 		# Necessary to empty tmp because otherwise previous embeddings will be included 
-		for dir_ in os.listdir('tmp/'):
-			shutil.rmtree(f'tmp/{dir_}')
+		if os.path.isdir('tmp/'):
+			for dir_ in os.listdir('tmp/'):
+				shutil.rmtree(f'tmp/{dir_}')
 		self.ann = AnnLite(self.rank, metric='cosine', data_path="tmp/")
 		docs = DocumentArray.empty(self.texel_embeddings.shape[0])
 		docs.embeddings = self.texel_embeddings
